+++ src/LIGHTActuator.h
@@ -1,6 +1,6 @@
 bool LIGHT_Setup();
 
-int LIGHT_GET_data();
+float LIGHT_GET_data();
 
 bool LIGHT_SENSOR_ANALOG_Setup();
 bool LIGHT_SENSOR_DIGITAL_Setup();
